requirements.txt
alembic==1.4.2
argh==0.26.2
blinker==1.4
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.1
dnspython==1.16.0
dominate==2.5.1
email-validator==1.0.5
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.10.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Mobility==1.0.0
Flask-Moment==0.9.0
Flask-Sitemap==0.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-core==1.6.0
google-cloud-logging==2.3.1
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.37.1
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==1.4.1
jwt==1.0.0
Mako==1.1.2
MarkupSafe==1.1.1
packaging==20.9
pilkit==2.0
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.16
twilio==6.58.0
Unidecode==1.1.1
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.0
zipp==3.1.0
